Recuperation pointApparition avec puzzle :: Reste maj carte + verif zone

+++ app/src/main/java/ccm/cours/nicolas/tiniki/Entity/Position.java
@@ -2,22 +2,22 @@ package ccm.cours.nicolas.tiniki.Entity;
 
 public class Position {
 
-    private Long latitude;
-    private Long longitude;
+    private double latitude;
+    private double longitude;
 
-    public Long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public Long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
